Backend/requirements.txt
uvicorn==0.25.0
fastapi==0.106.0
pydantic==2.5.3
pandas==2.1.4
scikit-learn==1.3.2
gemini-api==0.1.6
langchain-google-vertexai
python-dotenv==1.0.0